LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/scale-composition-expected.txt


FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (-0.5) should be [5 1] assert_equals: expected "5 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0) should be [6 1] assert_equals: expected "6 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0.25) should be [6.5 1] assert_equals: expected "6.5 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0.5) should be [7 1] assert_equals: expected "7 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0.75) should be [7.5 1] assert_equals: expected "7.5 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (1) should be [8 1] assert_equals: expected "8 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (1.5) should be [9 1] assert_equals: expected "9 1 " but got "2 1 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (-0.5) should be [2.5 11 22.5] assert_equals: expected "2.5 11 22.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0) should be [4 10 18] assert_equals: expected "4 10 18 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0.25) should be [4.75 9.5 15.75] assert_equals: expected "4.75 9.5 15.75 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0.5) should be [5.5 9 13.5] assert_equals: expected "5.5 9 13.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0.75) should be [6.25 8.5 11.25] assert_equals: expected "6.25 8.5 11.25 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (1) should be [7 8 9] assert_equals: expected "7 8 9 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (1.5) should be [8.5 7 4.5] assert_equals: expected "8.5 7 4.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (-1) should be [0.5 1] assert_equals: expected "0.5 1 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (0.25) should be [1.125 1] assert_equals: expected "1.13 1 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (0.75) should be [1.375 1] assert_equals: expected "1.38 1 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (1) should be [1.5 1] assert_equals: expected "1.5 1 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (2) should be [2 1] assert_equals: expected "2 1 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (-1) should be [7 9 11] assert_equals: expected "7 9 11 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (0) should be [4 5 6] assert_equals: expected "4 5 6 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (0.25) should be [3.25 4 4.75] assert_equals: expected "3.25 4 4.75 " but got "none "
FAIL Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (0.75) should be [1.75 2 2.25] assert_equals: expected "1.75 2 2.25 " but got "none "
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (1) should be [none]
FAIL Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (2) should be [-2 -3 -4] assert_equals: expected "- 2 - 3 - 4 " but got "none "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (-0.5) should be [-2 -1 0] assert_equals: expected "- 2 - 1 0 " but got "1 2 3 "
PASS Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0) should be [1 2 3]
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0.25) should be [2.5 3.5 4.5] assert_equals: expected "2.5 3.5 4.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0.5) should be [4 5 6] assert_equals: expected "4 5 6 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0.75) should be [5.5 6.5 7.5] assert_equals: expected "5.5 6.5 7.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (1) should be [7 8 9] assert_equals: expected "7 8 9 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (1.5) should be [10 11 12] assert_equals: expected "10 11 12 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (-0.5) should be [5.5 14.5 26.5] assert_equals: expected "5.5 14.5 26.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0) should be [4 10 18] assert_equals: expected "4 10 18 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0.25) should be [3.25 7.75 13.75] assert_equals: expected "3.25 7.75 13.75 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0.5) should be [2.5 5.5 9.5] assert_equals: expected "2.5 5.5 9.5 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0.75) should be [1.75 3.25 5.25] assert_equals: expected "1.75 3.25 5.25 " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (1) should be [none] assert_equals: expected "none " but got "1 2 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (1.5) should be [-0.5 -3.5 -7.5] assert_equals: expected "- 0.5 - 3.5 - 7.5 " but got "1 2 3 "
